Convert develop-static to TS
